Implement Interpret(), and impl. helper functions evalOperator, evalNumber
Interpret calls evalOperator or evalNumber based on token.tokenType.
If called on an empty tree, returns an error.

evalNumber returns the number stored by the token.

evalOperator calls Interpret on both child nodes of current node, when
these are evaluated and return a float64 it passes these values to the
appropriate function from mathfunc and returns the result from that
operation.
If there is an error in the mathfunc function or if the operator is
invalid it returns an error.